Corect some bugs of the bsplin but there are one still remaning

+++ src/Materials/bsplinepatch.h
@@ -13,7 +13,7 @@ public:
 
 
 private:
-    glm::vec3 p(int u, int v);
+    glm::vec3 p(float u, float v);
     int getMinU();
     int getMinV();
     int getMaxU();
